Switch to compile-time DI as documented
This doesn't compile, because the imports in the documentation are
wrong. Resolving this problem is treally hard because following the
compile errors leads the engineer down irrelevant rabbit holes.

The key to fixing the problem is removing the `play.api._` import, but
it's hard to figure this out because removing this import breaks other
things.